refactor: separate MultHandler class
This commit separates the `MultHandler` class from `tkinterboard.py` into its own file `multhandler.py`. This modularization makes the codebase cleaner and more organized, promoting better code maintainability.

Changes made in this commit:
- Moved the `MultHandler` class to the `src/interfaces/multhandler.py` file.
- Removed the `MultHandler` class definition from `tkinterboard.py`.

No functional changes are made in this commit; it is purely a code organization update.
